chore: eventHandlers.js를 TypeScript로 마이그레이션
eventHandlers.js를 TypeScript로 변환하여 이벤트 핸들러의 타입 안정성을 강화함:
- 이벤트 객체에 대한 명확한 타입 지정 (KeyboardEvent, SubmitEvent 등)
- querySelector 사용 시 null 체크 추가하여 안전성 확보
- form 이벤트 핸들러를 변수(formSubmitHandler)에 저장하여 이벤트 제거 가능하도록 수정
- form 요소 타입 검사 (HTMLFormElement) 추가하여 타입 오류 방지
- distance 값을 Number로 변환하여 데이터 타입 일관성 유지
